Add react-router and affiliate login page
