+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^7.1.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^6.24.1",
     "react-scripts": "3.4.1"
   },
   "scripts": {
