+++ src/App.css
@@ -261,6 +261,15 @@ hr {
   cursor: pointer;
 }
 
+.btn__w-100 {
+  min-width: 100%;
+}
+.btn__icon {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
 .button__primary {
   background-color: var(--theme-primary);
   color: #fff;
@@ -514,4 +523,111 @@ df-messenger-chat-bubble {
 .titlebar-wrapper {
   border-top-right-radius: 16px;
   border-top-left-radius: 16px;
+}
+
+
+/* AFFILIATE LOGIN */
+
+.login__wrapper {
+  padding: 20px;
+  max-width: 500px;
+  margin: 0 auto;
+  height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+}
+.login__container {
+  width: 85%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+.login__container > * {
+  margin-bottom: 20px;
+}
+.login__container > a {
+  padding-bottom: 10px;
+}
+.login__header {
+  margin-bottom: 30px;
+  border-bottom: 2px solid #d5d5d5;
+  padding-bottom: 20px;
+  width: 100%;
+  text-align: center;
+}
+.login__container .input {
+  padding: 10px;
+  width: 100%;
+  border: 1px solid #d5d5d5;
+  border-radius: 3px;
+  color: var(--theme-primary);
+}
+.line__through {
+  width: 100%;
+  text-align: center;
+  border-bottom: 1px solid #d5d5d5;
+  line-height: 0.1em;
+  margin-bottom: 40px;
+}
+.line__through span {
+  background:#fff;
+  color: var(--theme-text-primary-dark);
+  padding:0 10px;
+}
+.btn__container button {
+  margin-right: 10px;
+}
+.btn {
+  text-decoration: none;
+  text-align: center;
+  min-width: 130px;
+  height: 40px;
+  padding: 8px 12px 8px 12px;
+  font-weight: 600;
+  line-height: 19px;
+  border-radius: 4px;
+  font-size: 14px;
+  outline: none;
+  cursor: pointer;
+  text-transform: uppercase;
+  border: 2px solid var(--theme-primary);
+}
+.btn__facebook {
+  background-color: #4267B2;
+  border: 1px solid #4267B2;
+  color: #FFF;
+  text-transform: none;
+  font-weight: 400;
+  display: flex;
+  align-items: center;
+  text-decoration: none;
+  width: 100%;
+  border-radius: 25px;
+}
+.btn__facebook .text {
+  text-align: center;
+  flex-grow: 1;
+  margin-left: -16px;
+  margin-bottom: -4px;
+}
+.btn__google {
+  background-color: #FFF;
+  border: 1px solid #d5d5d5;
+  color: var(--theme-text-primary-dark);
+  text-transform: none;
+  font-weight: 600;
+  display: flex;
+  align-items: center;
+  text-decoration: none;
+  width: 100%;
+  border-radius: 25px;
+}
+.btn__google .text {
+  text-align: center;
+  flex-grow: 1;
+  margin-left: -16px;
+  margin-bottom: -4px;
 }
